moving the smart contracts code related to its own file
