+++ config.json
@@ -1,1 +1,19 @@
-{"chainId":"b8923f70-c1f7-497f-961j","rpcNodePort":5000,"dataDirectory":"./data/","databaseFilePath":"database.db","autosaveInterval":1800000,"javascriptVMTimeout":10000,"streamNodes":["https://api.steemit.com","https://rpc.buildteam.io","https://rpc.steemviz.com","https://steemd.minnowsupportproject.org"],"startSteemBlock":25647683,"genesisSteemBlock":25647669,"keyCertificate":"","certificate":"","chainCertificate":""}
+{
+    "chainId": "b8923f70-c1f7-497f-961j",
+    "rpcNodePort": 5000,
+    "dataDirectory": "./data/",
+    "databaseFilePath": "database.db",
+    "autosaveInterval": 1800000,
+    "javascriptVMTimeout": 10000,
+    "streamNodes": [
+        "https://api.steemit.com",
+        "https://rpc.buildteam.io",
+        "https://rpc.steemviz.com",
+        "https://steemd.minnowsupportproject.org"
+    ],
+    "startSteemBlock": 25647669,
+    "genesisSteemBlock": 25647669,
+    "keyCertificate": "",
+    "certificate": "",
+    "chainCertificate": ""
+}
